% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_y_axis_scaling.R
\name{ABSOLUTE_CHART_TYPES}
\alias{ABSOLUTE_CHART_TYPES}
\title{Chart types that use absolute internal unit (no scaling)}
\usage{
ABSOLUTE_CHART_TYPES
}
\description{
Chart types that use absolute internal unit (no scaling)
}
\keyword{internal}
