% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\name{AUTO_RESTORE_ENABLED}
\alias{AUTO_RESTORE_ENABLED}
\title{Auto restore funktion}
\usage{
AUTO_RESTORE_ENABLED
}
\description{
Bestemmer om session data automatisk skal gendannes ved app start.
}
\keyword{internal}
