% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_imports.R
\docType{package}
\name{biSPCharts-package}
\alias{biSPCharts}
\alias{biSPCharts-package}
\title{Package Imports}
\description{
En Shiny applikation til Statistical Process Control (SPC) analyse specifikt designet til brug i sundhedsvæsenet. Appen giver mulighed for at uploade data, konfigurere SPC grafer og eksportere resultater til kvalitetsarbejde i kliniske miljøer. Pakket som golem-baseret Shiny applikation med robust error handling og test coverage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/johanreventlow/biSPCharts}
  \item Report bugs at \url{https://github.com/johanreventlow/biSPCharts/issues}
}

}
\author{
\strong{Maintainer}: Johan Reventlow \email{johan@reventlow.dk}

}
\keyword{internal}
