% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spc_chart_ui_helpers.R
\name{build_antal_kryds_box}
\alias{build_antal_kryds_box}
\title{Build Antal Kryds Value Box}
\usage{
build_antal_kryds_box(status_info, anhoej)
}
\description{
Helper to build the crossing count (antal kryds) value box.
}
\keyword{internal}
