% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spc_chart_ui_helpers.R
\name{build_serielengde_box}
\alias{build_serielengde_box}
\title{Build Serielaengde Value Box}
\usage{
build_serielengde_box(status_info, anhoej)
}
\description{
Helper to build the longest run (serielaengde) value box.
}
\keyword{internal}
