% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spc_file_save_load.R
\name{build_spc_excel}
\alias{build_spc_excel}
\title{Byg biSPCharts Excel-fil med Data- og Indstillinger-ark}
\usage{
build_spc_excel(data, metadata)
}
\arguments{
\item{data}{data.frame med brugerens data}

\item{metadata}{Named list svarende til collect_metadata()-output}
}
\value{
Karakter-streng: sti til midlertidig .xlsx-fil
}
\description{
Byg biSPCharts Excel-fil med Data- og Indstillinger-ark
}
\keyword{internal}
