% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spc_cache.R
\name{clear_spc_cache}
\alias{clear_spc_cache}
\title{Clear All SPC Cache Entries}
\usage{
clear_spc_cache(cache)
}
\arguments{
\item{cache}{cache object. QIC cache instance fra \code{get_or_init_qic_cache()} eller \code{create_qic_cache()}.}
}
\value{
TRUE hvis ryddet succesfuldt, FALSE ved fejl.
}
\description{
Removes all entries from a QIC cache object. Wrapper around \code{cache$clear()}
som muliggør testbar og eksplicit oprydning af SPC-cachen.
}
\examples{
\dontrun{
qic_cache <- get_or_init_qic_cache(app_state)
clear_spc_cache(qic_cache)
}

}
\keyword{internal}
