% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_app_layout.R
\name{create_chart_settings_card_compact}
\alias{create_chart_settings_card_compact}
\title{Kompakt indstillings-card til hoejre side}
\usage{
create_chart_settings_card_compact()
}
\description{
Kompakt indstillings-card til hoejre side
}
\keyword{internal}
