% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_app_ui.R
\name{create_ui_main_content}
\alias{create_ui_main_content}
\title{Create UI main content area}
\usage{
create_ui_main_content()
}
\value{
tagList with main content components
}
\description{
Create UI main content area
}
\keyword{internal}
