% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\name{DEFAULT_ENCODING}
\alias{DEFAULT_ENCODING}
\title{Standard encoding for Windows kompatibilitet}
\usage{
DEFAULT_ENCODING
}
\description{
ISO-8859-1 encoding bruges for at sikre kompatibilitet med Windows systemer
og danske specialtegn i CSV filer.
}
\keyword{internal}
