% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\name{EXPORT_ASPECT_RATIO_MIN}
\alias{EXPORT_ASPECT_RATIO_MIN}
\title{Minimum aspect ratio for custom export dimensions}
\usage{
EXPORT_ASPECT_RATIO_MIN
}
\description{
Forhindrer alt for smalle plots (width/height < 0.5).
}
\keyword{internal}
