% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\name{EXPORT_DEPARTMENT_MAX_LENGTH}
\alias{EXPORT_DEPARTMENT_MAX_LENGTH}
\title{Maximum character length for department name}
\usage{
EXPORT_DEPARTMENT_MAX_LENGTH
}
\description{
Begraenser afdeling/afsnit navn til 250 karakterer.
}
\keyword{internal}
