% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\name{EXPORT_DESCRIPTION_MAX_LENGTH}
\alias{EXPORT_DESCRIPTION_MAX_LENGTH}
\title{Maximum character length for export description}
\usage{
EXPORT_DESCRIPTION_MAX_LENGTH
}
\description{
Begraenser beskrivelse laengde til 2000 karakterer for PDF metadata.
}
\keyword{internal}
