% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\name{EXPORT_FILENAME_PREFIX}
\alias{EXPORT_FILENAME_PREFIX}
\title{Default prefix for exported filenames}
\usage{
EXPORT_FILENAME_PREFIX
}
\description{
Alle eksporterede filer starter med denne prefix.
}
\keyword{internal}
