% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\name{EXPORT_FILENAME_SEPARATOR}
\alias{EXPORT_FILENAME_SEPARATOR}
\title{Separator character for filename components}
\usage{
EXPORT_FILENAME_SEPARATOR
}
\description{
Bruges til at separere prefix, titel og timestamp i filnavne.
}
\keyword{internal}
