% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\name{EXPORT_FORMAT_OPTIONS}
\alias{EXPORT_FORMAT_OPTIONS}
\title{Export format options}
\format{
Named character vector med format labels
}
\usage{
EXPORT_FORMAT_OPTIONS
}
\description{
Tilgaengelige eksport formater med display labels.
}
\keyword{internal}
