% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\docType{data}
\name{EXPORT_HOSPITAL_MAX_LENGTH}
\alias{EXPORT_HOSPITAL_MAX_LENGTH}
\title{Maximum character length for hospital name}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EXPORT_HOSPITAL_MAX_LENGTH
}
\description{
Begraenser hospitalsnavn til 250 karakterer.
}
\keyword{internal}
