% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\name{EXPORT_TITLE_MAX_LENGTH}
\alias{EXPORT_TITLE_MAX_LENGTH}
\title{Maximum character length for export title}
\usage{
EXPORT_TITLE_MAX_LENGTH
}
\description{
Begraenser titel laengde til 200 karakterer for at undgaa layout problemer.
}
\keyword{internal}
