% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analytics_pins.R
\name{filter_shinylogs_allowlist}
\alias{filter_shinylogs_allowlist}
\title{Filtrer shinylogs-data til allowlistede kolonner}
\usage{
filter_shinylogs_allowlist(all_data)
}
\arguments{
\item{all_data}{Navngivet liste fra \code{read_shinylogs_all()}}
}
\value{
Filtreret liste med samme struktur
}
\description{
Hash'er sessionid-kolonnen, redacterer fejlbeskeder og beholder
kun kolonner fra \code{SHINYLOGS_ALLOWLIST}.
}
\keyword{internal}
