% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_table_update_service.R
\name{format_data_for_excelr}
\alias{format_data_for_excelr}
\title{Formater data til excelR-visning}
\usage{
format_data_for_excelr(data)
}
\arguments{
\item{data}{Data frame der skal formateres}
}
\value{
Data frame med numeriske kolonner konverteret til character
}
\description{
Ren hjaelpefunktion der konverterer numeriske kolonner til dansk
format (komma-decimal) for visning i excelR-tabel.
Logiske kolonner (Skift/Frys) ekskluderes fra numerisk formatering.
}
\keyword{internal}
