% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export_download.R
\name{generate_pdf_export}
\alias{generate_pdf_export}
\title{Generate PDF Export}
\usage{
generate_pdf_export(input, app_state, file)
}
\arguments{
\item{input}{Shiny input object}

\item{app_state}{Global app state}

\item{file}{Output file path}
}
\description{
Generate PDF Export
}
\keyword{internal}
