% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export_download.R
\name{generate_png_export}
\alias{generate_png_export}
\title{Generate PNG Export}
\usage{
generate_png_export(input, app_state, file)
}
\arguments{
\item{input}{Shiny input object}

\item{app_state}{Global app state}

\item{file}{Output file path}
}
\description{
Generate PNG Export
}
\keyword{internal}
