% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export_download.R
\name{generate_pptx_export}
\alias{generate_pptx_export}
\title{Generate PowerPoint Export}
\usage{
generate_pptx_export(input, app_state, file)
}
\arguments{
\item{input}{Shiny input object}

\item{app_state}{Global app state}

\item{file}{Output file path}
}
\description{
Generate PowerPoint Export
}
\keyword{internal}
