% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_analytics.R
\name{get_analytics_config}
\alias{get_analytics_config}
\title{Hent analytics konfiguration}
\usage{
get_analytics_config()
}
\value{
List med analytics konfiguration
}
\description{
Returnerer analytics config. Kan udvides til at laese fra
golem-config.yml i fremtiden.
}
