% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_auto_save_enabled}
\alias{get_auto_save_enabled}
\title{Get auto save enabled setting}
\usage{
get_auto_save_enabled()
}
\value{
Boolean indicating if continuous auto-save is enabled
}
\description{
Get auto save enabled setting
}
\keyword{internal}
