% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{get_error_history}
\alias{get_error_history}
\title{Get Error History}
\usage{
get_error_history(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
List of error records (oldest first)
}
\description{
Returns the recent error history (max \code{ERROR_HISTORY_CAP}, FIFO).
}
\keyword{internal}
