% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{get_frys_column}
\alias{get_frys_column}
\title{Hent frys-kolonne mapping}
\usage{
get_frys_column(app_state)
}
\arguments{
\item{app_state}{Centraliseret app state}
}
\value{
Character eller NULL
}
\description{
Returnerer den aktuelle frys_column-mapping fra app_state.
}
\keyword{internal}
