% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{get_has_data_status}
\alias{get_has_data_status}
\alias{set_has_data_status}
\title{Get/Set Has-Data Status}
\usage{
get_has_data_status(app_state)

set_has_data_status(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{"true" eller "false"}
}
\description{
Wizard-gate-status: "true"/"false"-streng (renderText-kompatibel) der
signalerer om der er meningsfuld data tilgængelig.
}
\keyword{internal}
