% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{get_kommentar_column}
\alias{get_kommentar_column}
\title{Hent kommentar-kolonne mapping}
\usage{
get_kommentar_column(app_state)
}
\arguments{
\item{app_state}{Centraliseret app state}
}
\value{
Character eller NULL
}
\description{
Returnerer den aktuelle kommentar_column-mapping fra app_state.
}
\keyword{internal}
