% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{get_last_recovery_time}
\alias{get_last_recovery_time}
\title{Get Last Recovery Time}
\usage{
get_last_recovery_time(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
POSIXct or NULL
}
\description{
Get Last Recovery Time
}
\keyword{internal}
