% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{get_last_upload_time}
\alias{get_last_upload_time}
\alias{set_last_upload_time}
\title{Get/Set Last Upload Time}
\usage{
get_last_upload_time(app_state)

set_last_upload_time(app_state, value = Sys.time())
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{POSIXct timestamp (default \code{Sys.time()})}
}
\description{
Timestamp for seneste vellykket file-upload. Bruges af rate-limit-
check i fct_file_operations.
}
\keyword{internal}
