% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{get_n_column}
\alias{get_n_column}
\title{Hent N-kolonne (nævner) mapping}
\usage{
get_n_column(app_state)
}
\arguments{
\item{app_state}{Centraliseret app state}
}
\value{
Character eller NULL
}
\description{
Returnerer den aktuelle n_column-mapping fra app_state.
}
\keyword{internal}
