% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{get_recovery_attempts}
\alias{get_recovery_attempts}
\title{Get Recovery Attempts}
\usage{
get_recovery_attempts(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
Integer
}
\description{
Get Recovery Attempts
}
\keyword{internal}
