% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_settings_save_interval_ms}
\alias{get_settings_save_interval_ms}
\title{Get auto-save debounce interval for settings changes (milliseconds)}
\usage{
get_settings_save_interval_ms()
}
\value{
Integer milliseconds
}
\description{
Get auto-save debounce interval for settings changes (milliseconds)
}
\keyword{internal}
