% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_spc_config.R
\name{get_spc_hard_min}
\alias{get_spc_hard_min}
\title{Hent SPC hard minimum datapunkter}
\usage{
get_spc_hard_min()
}
\value{
Integer. Matematisk minimum for SPC-analyse.
}
\description{
Hent SPC hard minimum datapunkter
}
\keyword{internal}
