% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_spc_config.R
\name{get_spc_recommended_threshold}
\alias{get_spc_recommended_threshold}
\title{Hent SPC anbefalet minimum datapunkter}
\usage{
get_spc_recommended_threshold()
}
\value{
Integer. Anbefalet minimum for palidelig SPC-analyse.
}
\description{
Hent SPC anbefalet minimum datapunkter
}
\keyword{internal}
