% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_spc_config.R
\name{get_spc_warning_threshold}
\alias{get_spc_warning_threshold}
\title{Hent SPC advarselsgransens for kort serie}
\usage{
get_spc_warning_threshold()
}
\value{
Integer. Antal datapunkter under hvilke Anhoej-rules er upalidelige.
}
\description{
Hent SPC advarselsgransens for kort serie
}
\keyword{internal}
