% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{get_y_column}
\alias{get_y_column}
\title{Hent Y-kolonne mapping}
\usage{
get_y_column(app_state)
}
\arguments{
\item{app_state}{Centraliseret app state}
}
\value{
Character eller NULL
}
\description{
Returnerer den aktuelle y_column-mapping fra app_state.
}
\keyword{internal}
