% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{increment_recovery_attempts}
\alias{increment_recovery_attempts}
\title{Increment Recovery Attempts}
\usage{
increment_recovery_attempts(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\description{
Adds 1 to the recovery-attempt counter.
}
\keyword{internal}
