% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_startup_cache.R
\name{init_startup_cache}
\alias{init_startup_cache}
\title{Initialize startup cache directory}
\usage{
init_startup_cache()
}
\value{
TRUE if cache dir exists or was created, FALSE otherwise
}
\description{
Initialize startup cache directory
}
\keyword{internal}
