% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analytics_github.R
\name{inject_pat_into_url}
\alias{inject_pat_into_url}
\title{Indsæt PAT i HTTPS git-URL for autentificering}
\usage{
inject_pat_into_url(url, pat)
}
\arguments{
\item{url}{HTTPS git URL}

\item{pat}{GitHub Personal Access Token}
}
\value{
URL med indlejret PAT
}
\description{
Omsætter \verb{https://github.com/owner/repo.git} til
\verb{https://x-access-token:<PAT>@github.com/owner/repo.git}
som gert::git_clone/push kan bruge.
}
\keyword{internal}
