% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{is_autogen_active}
\alias{is_autogen_active}
\title{Get Autogen Active Flag}
\usage{
is_autogen_active(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
Logical (default FALSE)
}
\description{
Get Autogen Active Flag
}
\keyword{internal}
