% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{is_plot_generating}
\alias{is_plot_generating}
\title{Check if Plot is Generating}
\usage{
is_plot_generating(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
Logical
}
\description{
Safely checks if a plot generation is in progress.
}
\keyword{internal}
