% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_spc_config.R
\name{MAX_MISSING_PERCENT}
\alias{MAX_MISSING_PERCENT}
\title{Maximum missing values procent foer advarsel}
\usage{
MAX_MISSING_PERCENT
}
\description{
Maximum missing values procent foer advarsel
}
\keyword{internal}
