% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_spc_config.R
\name{MIN_NUMERIC_PERCENT}
\alias{MIN_NUMERIC_PERCENT}
\title{Minimum procent numeriske vaerdier for kolonne detection}
\usage{
MIN_NUMERIC_PERCENT
}
\description{
Minimum procent numeriske vaerdier for kolonne detection
}
\keyword{internal}
