% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_help_server.R
\name{mod_help_server}
\alias{mod_help_server}
\title{Help Module Server}
\usage{
mod_help_server(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
NULL (ingen outputs)
}
\description{
Minimal server logik for hjælpesiden.
Indholdet er statisk, så ingen reaktivitet er nødvendig.
}
