% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_help_ui.R
\name{mod_help_ui}
\alias{mod_help_ui}
\title{Help Module UI}
\usage{
mod_help_ui(id)
}
\arguments{
\item{id}{Character. Namespace ID for modulet}
}
\value{
Shiny UI element
}
\description{
Introduktion til SPC-begreber og praktisk brug i biSPCharts.
Indholdet er statisk og kraever ingen server-logik.
}
\keyword{internal}
