% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_landing_ui.R
\name{mod_landing_ui}
\alias{mod_landing_ui}
\title{Landing Page Module UI}
\usage{
mod_landing_ui(id)
}
\arguments{
\item{id}{Character. Namespace ID for modulet}
}
\value{
Shiny UI element
}
\description{
Velkomstside der renderes dynamisk baseret på localStorage-peek.
Når en gemt session er tilgængelig vises en restore-card; ellers
vises standard-landingen med "Kom i gang"-knap.
}
