% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\name{OPERATION_TIMEOUTS}
\alias{OPERATION_TIMEOUTS}
\title{Timeout værdier for forskellige operationer (millisekunder)}
\usage{
OPERATION_TIMEOUTS
}
\description{
Timeout værdier for forskellige operationer (millisekunder)
}
\keyword{internal}
