% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spc_file_save_load.R
\name{parse_spc_excel}
\alias{parse_spc_excel}
\title{Læs Indstillinger-ark fra biSPCharts Excel-fil}
\usage{
parse_spc_excel(file_path, sheets = NULL)
}
\arguments{
\item{file_path}{Sti til Excel-filen}
}
\value{
Named list svarende til collect_metadata()-output, eller NULL
hvis arket mangler eller er korrupt
}
\description{
Læs Indstillinger-ark fra biSPCharts Excel-fil
}
\keyword{internal}
