% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\name{PERFORMANCE_THRESHOLDS}
\alias{PERFORMANCE_THRESHOLDS}
\title{Performance monitoring thresholds}
\usage{
PERFORMANCE_THRESHOLDS
}
\description{
Performance monitoring thresholds
}
\keyword{internal}
