% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_autodetect_pure.R
\name{print.AutodetectResult}
\alias{print.AutodetectResult}
\title{Print-metode for AutodetectResult}
\usage{
\method{print}{AutodetectResult}(x, ...)
}
\arguments{
\item{x}{AutodetectResult-objekt.}

\item{...}{Ignoreres.}
}
\description{
Print-metode for AutodetectResult
}
