% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_file_parse_pure.R
\name{print.ParsedFile}
\alias{print.ParsedFile}
\title{Print-metode for ParsedFile}
\usage{
\method{print}{ParsedFile}(x, ...)
}
\arguments{
\item{x}{ParsedFile-objekt.}

\item{...}{Ignoreres.}
}
\description{
Print-metode for ParsedFile
}
