% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visualization_config_pure.R
\name{print.VisualizationConfig}
\alias{print.VisualizationConfig}
\title{Print-metode for VisualizationConfig}
\usage{
\method{print}{VisualizationConfig}(x, ...)
}
\arguments{
\item{x}{VisualizationConfig-objekt.}

\item{...}{Ignoreres.}
}
\description{
Print-metode for VisualizationConfig
}
