% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analytics_pins.R
\name{read_shinylogs_all}
\alias{read_shinylogs_all}
\title{Laes alle shinylogs kategorier til navngivet liste}
\usage{
read_shinylogs_all(log_directory)
}
\arguments{
\item{log_directory}{Sti til log-mappe}
}
\value{
Navngivet liste med 4 data.frames: sessions, inputs, outputs, errors
}
\description{
Laeser sessions, inputs, outputs og errors fra shinylogs log-directory.
}
