% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analytics_pins.R
\name{read_shinylogs_sessions}
\alias{read_shinylogs_sessions}
\title{Laes shinylogs session-filer til data.frame}
\usage{
read_shinylogs_sessions(log_directory)
}
\arguments{
\item{log_directory}{Sti til log-mappe}
}
\value{
data.frame med session-data (tom data.frame hvis ingen filer)
}
\description{
Laes shinylogs session-filer til data.frame
}
