% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_spc_config.R
\name{RECOMMENDED_SPC_POINTS}
\alias{RECOMMENDED_SPC_POINTS}
\title{Anbefalet minimum antal punkter for SPC}
\usage{
RECOMMENDED_SPC_POINTS
}
\description{
Anbefalet minimum antal punkter for SPC
}
\keyword{internal}
