% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_advanced_debug.R
\name{redact_col_names}
\alias{redact_col_names}
\title{Redaktér PII-kolonnenavne}
\usage{
redact_col_names(col_names)
}
\arguments{
\item{col_names}{Character-vector med kolonnenavne}
}
\value{
Character-vector, samme længde
}
\description{
Erstatter navne der matcher kendte PII-mønstre med \verb{[redacted]}.
}
\keyword{internal}
