% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analytics_pins.R
\name{rotate_log_files}
\alias{rotate_log_files}
\title{Roter log-filer (komprimer gamle, slet meget gamle)}
\usage{
rotate_log_files(
  log_directory,
  compress_after_days = ANALYTICS_CONFIG$log_compress_after_days,
  delete_after_days = ANALYTICS_CONFIG$log_retention_days
)
}
\arguments{
\item{log_directory}{Sti til log-mappe}

\item{compress_after_days}{Komprimer filer aeldre end dette (default: 90)}

\item{delete_after_days}{Slet filer aeldre end dette (default: 365)}
}
\description{
Roter log-filer (komprimer gamle, slet meget gamle)
}
\keyword{internal}
