% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_sample_data.R
\name{SAMPLE_DATASETS}
\alias{SAMPLE_DATASETS}
\title{Sample Dataset Definitions}
\format{
Named list of lists
}
\usage{
SAMPLE_DATASETS
}
\description{
Liste over eksempeldatasaet med metadata til dropdown-menu.
Hvert entry indeholder id, label, beskrivelse, filnavn og anbefalet chart type.
}
\keyword{internal}
