% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_current_data}
\alias{set_current_data}
\title{Set Current Data}
\usage{
set_current_data(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{Data frame to set}
}
\description{
Safely sets the current data in app_state.
}
\keyword{internal}
