% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_last_recovery_time}
\alias{set_last_recovery_time}
\title{Set Last Recovery Time}
\usage{
set_last_recovery_time(app_state, value = Sys.time())
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{POSIXct timestamp (default \code{Sys.time()})}
}
\description{
Set Last Recovery Time
}
\keyword{internal}
