% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_plot_generating}
\alias{set_plot_generating}
\title{Set Plot Generating Flag}
\usage{
set_plot_generating(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{Logical value}
}
\description{
Sets the plot-generating (is_computing) flag in app_state.
}
\keyword{internal}
