% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_plot_object}
\alias{set_plot_object}
\title{Set Plot Object}
\usage{
set_plot_object(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{ggplot object or NULL}
}
\description{
Sets the plot object in app_state.
}
\keyword{internal}
