% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_table_op_cleanup_needed}
\alias{set_table_op_cleanup_needed}
\title{Set Table Operation Cleanup Needed Flag}
\usage{
set_table_op_cleanup_needed(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{Logical TRUE/FALSE}
}
\description{
Markerer at en table-opdatering kræver post-flush cleanup (clearing
table_operation_in_progress flag). Forhindrer race conditions
mellem rapid table-edits.
}
\keyword{internal}
