% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_ui_updating}
\alias{set_ui_updating}
\title{Set UI Updating Programmatically Flag}
\usage{
set_ui_updating(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{Logical TRUE/FALSE}
}
\description{
Toggle for igangvaerende programmatisk UI-update. Brugt af
safe_programmatic_ui_update() til race-protection mod brugerinteraktion.
}
\keyword{internal}
