% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_viz_cache_updating}
\alias{set_viz_cache_updating}
\title{Set Visualization Cache Updating Flag}
\usage{
set_viz_cache_updating(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{Logical TRUE/FALSE}
}
\description{
Toggle for igangvaerende visualization-cache-opdatering. Forhindrer
samtidige cache-rebuilds (mod_spc_chart_state).
}
\keyword{internal}
