% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analytics_consent.R
\name{setup_analytics_consent}
\alias{setup_analytics_consent}
\title{Setup analytics consent observer i Shiny server}
\usage{
setup_analytics_consent(input, session, hashed_token, log_directory = "logs/")
}
\arguments{
\item{input}{Shiny input object}

\item{session}{Shiny session object}

\item{hashed_token}{Hashed session token for logging}

\item{log_directory}{Directory for shinylogs output}
}
\value{
Invisible NULL
}
\description{
Registrerer observer paa input$analytics_consent og initialiserer
shinylogs KUN naar brugeren har accepteret.
}
