% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server_event_listeners.R
\name{setup_paste_data_observers}
\alias{setup_paste_data_observers}
\title{Setup observers for paste data og sample data loading}
\usage{
setup_paste_data_observers(input, app_state, session, emit)
}
\arguments{
\item{input}{Shiny input}

\item{app_state}{Centraliseret app state}

\item{session}{Shiny session}

\item{emit}{Event emit API}
}
\description{
Setup observers for paste data og sample data loading
}
\keyword{internal}
