% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server_wizard_gates.R
\name{setup_wizard_gates}
\alias{setup_wizard_gates}
\title{Setup wizard navigation gates}
\usage{
setup_wizard_gates(input, output, app_state, session, emit)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{app_state}{Centraliseret app state}

\item{session}{Shiny session}

\item{emit}{Event emission API}
}
\description{
Locks/unlocks navbar wizard steps based on app state.
Trin 1 (Upload) altid tilgaengelig. Trin 2 (Analyser) kraever data.
Trin 3 (Eksporter) kraever renderet plot.
}
\keyword{internal}
