% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_spc_config.R
\name{SPC_DATA_THRESHOLDS}
\alias{SPC_DATA_THRESHOLDS}
\title{SPC datapunkt-taerskler}
\usage{
SPC_DATA_THRESHOLDS
}
\description{
Samlet liste med alle datapunkt-graenser brugt paa tvaers af SPC-pipelinen.
hard_min       = matematisk minimum for SPC-beregning
warning_short  = under dette niveau er Anhoej-rules upalidelige
recommended    = konventionel SPC best-practice (mindst 15 punkter)
}
\keyword{internal}
