% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time_formatting.R
\docType{data}
\name{TIME_BREAK_CANDIDATES}
\alias{TIME_BREAK_CANDIDATES}
\title{Tids-naturlige kandidat-intervaller i minutter}
\format{
An object of class \code{numeric} of length 17.
}
\usage{
TIME_BREAK_CANDIDATES
}
\description{
Bruges af time_breaks() til at vaelge tick-afstand. Daekker fra 1 minut
op til 30 dage.
}
\keyword{internal}
