% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time_formatting.R
\docType{data}
\name{TIME_BREAK_CANDIDATES_BY_UNIT}
\alias{TIME_BREAK_CANDIDATES_BY_UNIT}
\title{Tilladte kandidat-intervaller pr. input-enhed}
\format{
An object of class \code{list} of length 3.
}
\usage{
TIME_BREAK_CANDIDATES_BY_UNIT
}
\description{
Filtrerer TIME_BREAK_CANDIDATES baseret paa hvilke der er "naturlige"
i den valgte input-enhed. Fx time_hours tillader ikke 15m-intervaller
(= 0.25 t) men tillader 30m (= 0.5 t). time_days kraever mindst 12t
interval for at undgaa forvirrende sub-dag ticks.
}
\keyword{internal}
