% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_ui.R
\name{UNIT_TYPE_LABELS}
\alias{UNIT_TYPE_LABELS}
\title{Kliniske enhedstype-labels til UI-selector}
\usage{
UNIT_TYPE_LABELS
}
\description{
Bruges i \code{current_unit()} til at mappe korte enhedskoder til
fulde danske navne. Tilfoej nye enhedstyper her.
}
\keyword{internal}
