% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\docType{data}
\name{UPLOAD_LIMITS}
\alias{UPLOAD_LIMITS}
\title{Upload validation thresholds}
\format{
An object of class \code{list} of length 4.
}
\usage{
UPLOAD_LIMITS
}
\description{
Upload validation thresholds
}
\keyword{internal}
