% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\name{UTF8_ENCODING}
\alias{UTF8_ENCODING}
\title{Alternative encoding for UTF-8 filer}
\usage{
UTF8_ENCODING
}
\description{
UTF-8 encoding til moderne tekst filer og internationale tegnsæt.
}
\keyword{internal}
