% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spc_contracts.R
\name{validate_spc_prepared_contract}
\alias{validate_spc_prepared_contract}
\title{Valider at et objekt overholder spc_prepared-kontrakten}
\usage{
validate_spc_prepared_contract(x)
}
\arguments{
\item{x}{Objekt der skal valideres.}
}
\value{
Usynlig \code{x} hvis gyldig; kaster fejl ellers.
}
\description{
Valider at et objekt overholder spc_prepared-kontrakten
}
\keyword{internal}
