% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spc_contracts.R
\name{validate_spc_request_contract}
\alias{validate_spc_request_contract}
\title{Valider at et objekt overholder spc_request-kontrakten}
\usage{
validate_spc_request_contract(x)
}
\arguments{
\item{x}{Objekt der skal valideres.}
}
\value{
Usynlig \code{x} hvis gyldig; kaster fejl ellers.
}
\description{
Bruges i tests. Kaster informative fejl ved brud.
}
\keyword{internal}
