% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analytics_pins.R
\name{aggregate_and_pin_logs}
\alias{aggregate_and_pin_logs}
\title{Aggreger logs og publicer til Connect Cloud via pins}
\usage{
aggregate_and_pin_logs(log_directory = "logs/")
}
\arguments{
\item{log_directory}{Sti til log-mappe}
}
\description{
Aggreger logs og publicer til Connect Cloud via pins
}
