% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spc_bfh_facade.R
\name{build_cache_key}
\alias{build_cache_key}
\title{Byg cache-nøgle for SPC-beregning}
\usage{
build_cache_key(
  data,
  chart_type,
  x_var,
  y_var,
  n_var,
  multiply,
  extra_params,
  use_cache
)
}
\description{
Byg cache-nøgle for SPC-beregning
}
\keyword{internal}
