% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\name{EXPORT_HOSPITAL_MAX_LENGTH}
\alias{EXPORT_HOSPITAL_MAX_LENGTH}
\title{Maximum character length for hospital name}
\usage{
EXPORT_HOSPITAL_MAX_LENGTH
}
\description{
Begraenser hospitalsnavn til 250 karakterer.
}
\keyword{internal}
