% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spc_file_save_load.R
\name{fct_spc_file_save_load}
\alias{fct_spc_file_save_load}
\title{Fil-baseret gem og indlæs}
\description{
To rene funktioner til at skrive og læse biSPCharts Excel-filer.
Filen har to ark: "Data" (de rå rækker) og "Indstillinger" (alle
UI-indstillinger som Felt/Værdi-tabel), svarende til hvad
collect_metadata() returnerer.
}
