% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server_export.R
\name{inject_template_assets}
\alias{inject_template_assets}
\title{Inject biSPCharts Template Assets into Export Temp Directory}
\usage{
inject_template_assets(template_dir)
}
\arguments{
\item{template_dir}{Sti til bfh-template mappen i temp directory}
}
\value{
invisible(TRUE) ved succes, invisible(FALSE) ved fejl
}
\description{
Kopierer fonts og images fra biSPCharts' template til en temp-mappe
hvor BFHcharts har oprettet Typst-template strukturen.
Bruges fordi BFHcharts' GitHub repo ikke inkluderer fonts/images.
}
\keyword{internal}
