% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_app_guide_ui.R
\name{mod_app_guide_ui}
\alias{mod_app_guide_ui}
\title{App Guide Module UI}
\usage{
mod_app_guide_ui(id)
}
\arguments{
\item{id}{Character. Namespace ID for modulet}
}
\value{
Shiny UI element
}
\description{
Detaljeret trin-for-trin vejledning i brug af biSPCharts-appen.
Indholdet er statisk og kr\u00e6ver ingen server-logik.
}
