% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_help_server.R
\name{mod_help_server}
\alias{mod_help_server}
\title{Help Module Server}
\usage{
mod_help_server(id, parent_session = NULL, previous_tab = NULL)
}
\arguments{
\item{id}{Module ID}

\item{parent_session}{Shiny session. Parent session for navbar navigation.}

\item{previous_tab}{ReactiveVal. Den forrige tab brugeren var på.}
}
\description{
Server logik for hjælpesiden. Håndterer tilbagenavigation
til den tab brugeren kom fra.
}
\keyword{internal}
