% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time_parsing.R
\name{TIME_INPUT_UNIT_SCALES}
\alias{TIME_INPUT_UNIT_SCALES}
\title{Skaleringsfaktor fra input-enhed til kanoniske minutter}
\usage{
TIME_INPUT_UNIT_SCALES
}
\description{
Skaleringsfaktor fra input-enhed til kanoniske minutter
}
\keyword{internal}
