% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_ui.R
\name{UI_HEIGHTS}
\alias{UI_HEIGHTS}
\title{Standard hoejder for UI komponenter}
\format{
Named list med CSS hoejde strings
}
\usage{
UI_HEIGHTS
}
\description{
CSS hoejde vaerdier til konsistent UI layout paa tvaers af komponenter.
}
\keyword{internal}
